Refactor structure again.
