+++ src/test/java/cloud/filibuster/functional/java_hello/JUnitFilibusterHelloPartialHelloWithErrorHandlingTest.java
@@ -1,4 +1,4 @@
-package cloud.filibuster.functional.java.hello.nondeterministic;
+package cloud.filibuster.functional.java_hello;
 
 import cloud.filibuster.examples.Hello;
 import cloud.filibuster.examples.HelloServiceGrpc;
@@ -19,18 +19,15 @@ import java.util.Set;
 import java.util.concurrent.TimeUnit;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * Test simple annotation usage.
  */
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitFilibusterNoFaultsNoTerminationTest extends JUnitBaseTest {
+public class JUnitFilibusterHelloPartialHelloWithErrorHandlingTest extends JUnitBaseTest {
     private final static Set<String> testExceptionsThrown = new HashSet<>();
 
     private static int numberOfTestsExecuted = 0;
-
-    private static int numberOfExceptionsThrown = 0;
 
     @DisplayName("Test partial hello server grpc route with Filibuster. (MyHelloService, MyWorldService)")
     @FilibusterTest(serverBackend=FilibusterLocalServerBackend.class, maxIterations=10)
@@ -43,16 +40,10 @@ public class JUnitFilibusterNoFaultsNoTerminationTest extends JUnitBaseTest {
 
         numberOfTestsExecuted++;
 
-        try {
-            HelloServiceGrpc.HelloServiceBlockingStub blockingStub = HelloServiceGrpc.newBlockingStub(helloChannel);
-            Hello.HelloRequest request = Hello.HelloRequest.newBuilder().setName("Armerian " + Math.random()).build();
-            Hello.HelloReply reply = blockingStub.partialHello(request);
-            assertTrue(reply.getMessage().contains("Hello, Armerian"));
-        } catch (RuntimeException e) {
-            numberOfExceptionsThrown++;
-            // Shouldn't ever get here.
-        }
-
+        HelloServiceGrpc.HelloServiceBlockingStub blockingStub = HelloServiceGrpc.newBlockingStub(helloChannel);
+        Hello.HelloRequest request = Hello.HelloRequest.newBuilder().setName("Armerian").build();
+        Hello.HelloReply reply = blockingStub.partialHelloWithErrorHandling(request);
+        assertEquals("Hello, Armerian World!!", reply.getMessage());
 
         helloChannel.shutdownNow();
         helloChannel.awaitTermination(1000, TimeUnit.SECONDS);
@@ -62,7 +53,6 @@ public class JUnitFilibusterNoFaultsNoTerminationTest extends JUnitBaseTest {
     @Test
     @Order(2)
     public void testNumAssertions() {
-        // No fault injections because of no DEI matches.
         assertEquals(0, testExceptionsThrown.size());
     }
 
@@ -70,15 +60,6 @@ public class JUnitFilibusterNoFaultsNoTerminationTest extends JUnitBaseTest {
     @Test
     @Order(3)
     public void testNumberOfTestsExecuted() {
-        // maxIterations executed because of no termination.
-        assertEquals(10, numberOfTestsExecuted);
-    }
-
-    @DisplayName("Verify correct number of exceptions thrown.")
-    @Test
-    @Order(4)
-    public void numberOfExceptionsThrown() {
-        // No fault injections because of no DEI matches.
-        assertEquals(0, numberOfExceptionsThrown);
+        assertEquals(5, numberOfTestsExecuted);
     }
 }
